added checks for passphrase in unattended mode
When using ZFS pool for root filesystem, the kernel modules must be
loaded before the availability of native encryption can be verified.

If available, a passphrase must be provided when  unattended mode.

Right after reading command line arguments, the kernel modules might
not be compiled/loaded yet, and it is not possible to verify whether
ZFS native encryption is supported. A passphrase is only neccessary in
unattended mode when native encryption is supported and not using a
LUKS root device.

Passphrase must be specified in unattended mode when:

1) When using LUKS as root device.

2) When using ZFS as root device, and the ZFS kernel modules have
been already loaded, and ZFS native encryption support is available,
and not explicitly disabled (option --without-zfs-native-encryption).

If ZFS kernel modules weren't available at start, but are supposed
to be compiled/loaded later on, then these additional scenarios must
be checked:

- when initialising a new ZFS pool
- when using ZFS for system root dataset
- when using LUKS encrypted root device, with additional ZFS datasets
to be mounted during boot